feat: support Azure federated identity for member-to-hub authentication
Replaces the azure token provider's Managed-Identity-only credential
(azidentity.NewManagedIdentityCredential) with
azidentity.NewDefaultAzureCredential, which chains through
EnvironmentCredential, WorkloadIdentityCredential, ManagedIdentityCredential,
AzureCLICredential, and AzureDeveloperCLICredential in order. This lets the
member agent authenticate to an AKS hub cluster via Azure Workload Identity
Federation, a service principal, or CLI-based login, in addition to managed
identity, without an operator having to pick a mode up front or KubeFleet
needing a dedicated provider per mechanism.

- pkg/authtoken/providers/azure: rewrite the provider around
  DefaultAzureCredential; fix the default scope to include the /.default
  suffix required by the EnvironmentCredential/WorkloadIdentityCredential
  legs (AADSTS1002012); drop the client-ID option entirely, since every
  legitimate way of steering the chain to a specific identity already sets
  AZURE_CLIENT_ID itself (the workload-identity webhook, an
  EnvironmentCredential secret) and a separate flag would only risk
  silently overriding whichever of those actually applies.
- cmd/authtoken: drop the now-removed --clientid flag.

Note: charts/member-agent's values.yaml still defaults azure.clientid to a
placeholder string, which the chart's generic provider-arg-forwarding loop
would pass as --clientid to this new binary. A follow-up PR will update the
chart; until then, config.provider: azure via the chart's own defaults
needs a values override to omit that key.

Verified end to end against a live pair of AKS clusters (Azure
AD-integrated hub, OIDC issuer + workload identity enabled member): the
member agent authenticated via a federated-credential token exchange with
no static secret, and separately via the member cluster's own kubelet
managed identity, both reaching MemberCluster Joined=True.

Fixes #865
