+++ cmd/authtoken/main.go
@@ -59,21 +59,18 @@ func parseArgs() (authtoken.Provider, error) {
 	secretCmd.Flags().StringVar(&secretNamespace, "namespace", "default", "Secret namespace (required)")
 	_ = secretCmd.MarkFlagRequired("namespace")
 
-	var clientID string
 	var scope string
 	azureCmd := &cobra.Command{
 		Use:  "azure",
 		Args: cobra.NoArgs,
 		Run: func(_ *cobra.Command, args []string) {
-			tokenProvider = azure.New(clientID, scope)
+			tokenProvider = azure.New(scope)
 		},
 	}
 
-	azureCmd.Flags().StringVar(&clientID, "clientid", "", "Azure AAD client ID (required)")
 	// TODO: this scope argument is specific for Azure provider. We should allow registering and parsing provider specific argument
 	// in provider level, instead of global level.
 	azureCmd.Flags().StringVar(&scope, "scope", "", "Azure AAD token scope (optional)")
-	_ = azureCmd.MarkFlagRequired("clientid")
 
 	rootCmd.AddCommand(secretCmd, azureCmd)
 	err = rootCmd.Execute()
